Fix mysqli_connect to database from web server on linux container hosts
As containers are native processes, utilization of the mechanisms for how the container host resolves localhost results in resolution to the Unix socket '/var/run/mysqld/mysqld.sock' during container deployment on Unix hosts. If PHP defaults to TCP and we only have connections permitted for apache user via localhost, the absence of definition of '127.0.0.1' for 'apache' user will result in connection refusal.